# server/src/middleware/__generated__/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"GoogleOAuthUser": {
				"properties": {
					"uid": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"photoURL": {
						"type": "string"
					}
				},
				"required": [
					"uid",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoogleOAuthResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/GoogleOAuthUser"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GoogleOAuthRequest": {
				"properties": {
					"idToken": {
						"type": "string"
					}
				},
				"required": [
					"idToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendVerificationCodeResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendVerificationCodeRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"purpose": {
						"type": "string",
						"enum": [
							"signup",
							"forgot-password"
						]
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyCodeResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyCodeRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"inputCode": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"inputCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyTokenResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyTokenRequest": {
				"properties": {
					"idToken": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"university": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"required": [
					"idToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthDomainAllowedRoles": {
				"items": {
					"type": "string",
					"enum": [
						"staff",
						"student"
					]
				},
				"type": "array"
			},
			"AuthDomain": {
				"properties": {
					"emailDomains": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"institutionName": {
						"type": "string"
					},
					"allowedRoles": {
						"$ref": "#/components/schemas/AuthDomainAllowedRoles"
					}
				},
				"required": [
					"emailDomains",
					"institutionName",
					"allowedRoles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddSignupEmailDomainRequest": {
				"properties": {
					"institutionName": {
						"type": "string"
					},
					"emailDomains": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"allowedRoles": {
						"$ref": "#/components/schemas/AuthDomainAllowedRoles"
					}
				},
				"required": [
					"institutionName",
					"emailDomains",
					"allowedRoles"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "enter 'Bearer {Firebase ID token}'"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/auth/google/verify": {
			"post": {
				"operationId": "VerifyToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GoogleOAuthResponse"
								}
							}
						}
					}
				},
				"description": "Verifies Google ID token and authenticates user.",
				"tags": [
					"Google OAuth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GoogleOAuthRequest"
							}
						}
					}
				}
			}
		},
...
		"/auth/send-verification-code": {
			"post": {
				"operationId": "SendVerificationCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SendVerificationCodeResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendVerificationCodeRequest"
							}
						}
					}
				}
			}
		},
		"/auth/verify-code": {
			"post": {
				"operationId": "VerifyCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyCodeResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyCodeRequest"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResetPasswordResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/auth/verify-token": {
			"post": {
				"operationId": "VerifyToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyTokenResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VerifyTokenRequest"
							}
						}
					}
				}
			}
		},
...
		"/auth/{domain_id}": {
			"delete": {
				"operationId": "RemoveValidSignupEmailDomain",
				"responses": {
					"200": {
						"description": "Valid signup domain removed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthDomain"
								}
							}
						}
					}
				},
				"description": "Allows admins to remove a valid signup email domain",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- the ID of the domain to delete",
						"in": "path",
						"name": "domain_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth": {
			"post": {
				"operationId": "AddValidSignupEmailDomain",
				"responses": {
					"201": {
						"description": "Valid signup email domain added successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthDomain"
								}
							}
						}
					}
				},
				"description": "Allows admins to add a new valid signup email domain",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddSignupEmailDomainRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}